Merge the validated CI repair: exclude intentionally malformed fixtures from ordinary test/lint collection, add required validation/type dependencies, repair typed parser/runtime boundaries, remove the dead duplicate runtime, and restore the full Python 3.9-3.13 test/build matrix. No ontology, qualification, or readiness promotion is implied.

+++ src/shannon_insight/cross_layer/conway.py
@@ -50,4 +50,4 @@ def conway_alignment(tensor: RelationTensor, t: int = -1) -> float:
     if total_imports == 0:
         return 1.0
 
-    return 1 - len(violations) / total_imports
+    return float(1 - len(violations) / total_imports)
